kycverfication/kyc_verification_backend/services/image_verification.py
# services/image_verification.py

import cv2
from PIL import Image

def verify_images(live_image_path, id_document_path):
    """Compare the live image with the ID document image."""
    # Load images using OpenCV
    live_image = cv2.imread(live_image_path)
    id_image = cv2.imread(id_document_path)

    # Implement a placeholder comparison logic
    # This can be replaced with a more advanced facial recognition library
    if live_image is None or id_image is None:
        return False, "Failed to load images."

    # Simple size comparison for demonstration (replace with facial recognition)
    if live_image.shape != id_image.shape:
        return False, "Images do not match."

    return True, "Images verified successfully."
